fix: fixed bug with incorrect path to features folder
